requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
boto3==1.24.51
botocore==1.27.51
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
cloudinary==1.29.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.5.0
django-storages==1.13.1
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.1
mock==4.0.3
msgpack==1.0.4
Pillow==9.2.0
psycopg2==2.9.3; sys_platform != 'linux'
psycopg2-binary>=2.8; sys_platform == 'linux'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.2
requests==2.28.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
zope.interface==5.4.0
